Source commit: 6f9ef689e119a87ec12a16671716367a4d79e0ad

+++ libs/cleaninterm.bat
@@ -1,34 +1,63 @@
 @echo off
 
+setlocal
+
+set ANY_CLEANED=0
+
 rem ==== OpenSSL ====
 
-echo Cleaning up OpenSSL ...
-rmdir /s /q openssl\tmp
+if exist openssl\tmp (
+    echo Cleaning up OpenSSL ...
+    rmdir /s /q openssl\tmp
+    set ANY_CLEANED=1
+)
 
 rem ==== Expat ====
 
-echo Cleaning up Expat ...
-rmdir /s /q expat\bcb5\release
+if exist expat\bcb5\release (
+    echo Cleaning up Expat ...
+    rmdir /s /q expat\bcb5\release
+    set ANY_CLEANED=1
+)
 
 rem ==== neon ====
 
-echo Cleaning up neon ...
-rmdir /s /q neon\tmp
+if exist neon\tmp (
+    echo Cleaning up neon ...
+    rmdir /s /q neon\tmp
+    set ANY_CLEANED=1
+)
 
 rem ==== PuTTYVS ====
 
-echo Cleaning up PuTTYVS ...
-rmdir /s /q puttyvs\Release
+if exist puttyvs\PuTTYVS (
+    echo Cleaning up PuTTYVS ...
+    rmdir /s /q puttyvs\PuTTYVS
+    set ANY_CLEANED=1
+)
 
 rem ==== libs3 ====
 
-echo Cleaning up libs3 ...
-rmdir /s /q libs3\tmp
+if exist libs3\tmp (
+    echo Cleaning up libs3 ...
+    rmdir /s /q libs3\tmp
+    set ANY_CLEANED=1
+)
 
 rem ==== MFC ====
 
-echo Cleaning up MFC ...
-rmdir /s /q mfc\source\$UW
-del mfc\source\bor.rsp
+set CLEAN_MFC=0
+if exist mfc\source\$UW set CLEAN_MFC=1
+if exist mfc\source\bor.rsp set CLEAN_MFC=1
+if %CLEAN_MFC% == 1 (
+    echo Cleaning up MFC ...
+    if exist mfc\source\$UW rmdir /s /q mfc\source\$UW
+    if exist mfc\source\bor.rsp del mfc\source\bor.rsp
+    set ANY_CLEANED=1
+)
 
-echo All done
+if %ANY_CLEANED% == 1 (
+    echo All done
+) else (
+    echo No clean needed
+)
